improve channel creation
